@@ -147,7 +147,9 @@ def test_direct_paging_user(make_organization, make_user_for_organization):
     assert alert.message == "Fire"
     # notifications sent
     for u, important in ((user, False), (other_user, True)):
-        assert notify_task.apply_async.called_with((u.pk, ag.pk), {"important": important})
+        assert notify_task.apply_async.called_with(
+            (u.pk, ag.pk), {"important": important, "notify_even_acknowledged": True, "notify_anyway": True}
+        )
         expected_info = {"user": u.public_primary_key, "schedule": None, "important": important}
         assert_log_record(ag, f"{from_user.username} paged user {u.username}", expected_info=expected_info)
 
@@ -182,7 +184,9 @@ def test_direct_paging_schedule(
     assert_log_record(ag, f"{from_user.username} paged schedule {other_schedule.name}")
     # notifications sent
     for u, important, s in ((user, False, schedule), (other_user, True, other_schedule)):
-        assert notify_task.apply_async.called_with((u.pk, ag.pk), {"important": important})
+        assert notify_task.apply_async.called_with(
+            (u.pk, ag.pk), {"important": important, "notify_even_acknowledged": True, "notify_anyway": True}
+        )
         expected_info = {"user": u.public_primary_key, "schedule": s.public_primary_key, "important": important}
         assert_log_record(
             ag, f"{from_user.username} paged user {u.username} (from schedule {s.name})", expected_info=expected_info
@@ -208,7 +212,9 @@ def test_direct_paging_reusing_alert_group(
     assert_log_record(alert_group, f"{from_user.username} paged user {user.username}")
     # notifications sent
     ag = alert_groups.get()
-    assert notify_task.apply_async.called_with((user.pk, ag.pk), {"important": False})
+    assert notify_task.apply_async.called_with(
+        (user.pk, ag.pk), {"important": False, "notify_even_acknowledged": True, "notify_anyway": True}
+    )
 
 
 @pytest.mark.django_db
@@ -308,5 +314,9 @@ def test_direct_paging_always_create_group(make_organization, make_user_for_orga
     alert_groups = AlertGroup.objects.all()
     assert alert_groups.count() == 2
     # notifications sent
-    assert notify_task.apply_async.called_with((user.pk, alert_groups[0].pk), {"important": False})
-    assert notify_task.apply_async.called_with((user.pk, alert_groups[1].pk), {"important": False})
+    assert notify_task.apply_async.called_with(
+        (user.pk, alert_groups[0].pk), {"important": False, "notify_even_acknowledged": True, "notify_anyway": True}
+    )
+    assert notify_task.apply_async.called_with(
+        (user.pk, alert_groups[1].pk), {"important": False, "notify_even_acknowledged": True, "notify_anyway": True}
+    )
