@@ -3,10 +3,11 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from apps.alerts.paging import direct_paging
+from apps.alerts.paging import DirectPagingAlertGroupResolvedError, direct_paging
 from apps.api.permissions import RBACPermission
 from apps.api.serializers.paging import DirectPagingSerializer
 from apps.auth_token.auth import PluginAuthentication
+from common.api_helpers.exceptions import BadRequest
 
 
 class DirectPagingAPIView(APIView):
@@ -30,17 +31,19 @@ class DirectPagingAPIView(APIView):
         schedules = [
             (schedule["instance"], schedule["important"]) for schedule in serializer.validated_data["schedules"]
         ]
-
-        alert_group = direct_paging(
-            organization=organization,
-            team=serializer.validated_data["team"],
-            from_user=from_user,
-            title=serializer.validated_data["title"],
-            message=serializer.validated_data["message"],
-            users=users,
-            schedules=schedules,
-            escalation_chain=serializer.validated_data["escalation_chain"],
-            alert_group=serializer.validated_data["alert_group"],
-        )
+        try:
+            alert_group = direct_paging(
+                organization=organization,
+                team=serializer.validated_data["team"],
+                from_user=from_user,
+                title=serializer.validated_data["title"],
+                message=serializer.validated_data["message"],
+                users=users,
+                schedules=schedules,
+                escalation_chain=serializer.validated_data["escalation_chain"],
+                alert_group=serializer.validated_data["alert_group"],
+            )
+        except DirectPagingAlertGroupResolvedError:
+            raise BadRequest(detail=DirectPagingAlertGroupResolvedError.DETAIL)
 
         return Response(data={"alert_group_id": alert_group.public_primary_key}, status=status.HTTP_200_OK)
