@@ -26,7 +26,7 @@ class SelfHostedInstallView(GrafanaHeadersMixin, APIView):
         provisioning_info = {"error": None}
 
         if settings.LICENSE != settings.OPEN_SOURCE_LICENSE_NAME:
-            provisioning_info["error"] = f"License type not authorized"
+            provisioning_info["error"] = "License type not authorized"
             return Response(status=status.HTTP_403_FORBIDDEN)
 
         grafana_api_client = GrafanaAPIClient(api_url=grafana_url, api_token=grafana_api_token)
