@@ -120,10 +120,10 @@ class PhoneBackend:
         if response.status_code == 200:
             logger.info("PhoneBackend._notify_by_cloud_call: OK")
         elif response.status_code == 400 and response.json().get("error") == "limit-exceeded":
-            logger.info(f"PhoneBackend._notify_by_cloud_call: phone calls limit exceeded")
+            logger.info("PhoneBackend._notify_by_cloud_call: phone calls limit exceeded")
             raise CallsLimitExceeded
         elif response.status_code == 400 and response.json().get("error") == "number-not-verified":
-            logger.info(f"PhoneBackend._notify_by_cloud_call: cloud number not verified")
+            logger.info("PhoneBackend._notify_by_cloud_call: cloud number not verified")
             raise NumberNotVerified
         elif response.status_code == 404:
             logger.info(f"PhoneBackend._notify_by_cloud_call: user not found id={user.id} email={user.email}")
