@@ -169,13 +169,13 @@ class ChannelFilterSerializer(BaseChannelFilterSerializer):
             try:
                 valid_jinja_template_for_serializer_method_field({"route_template": filtering_term})
             except JinjaTemplateError:
-                raise serializers.ValidationError([f"Jinja template is incorrect"])
+                raise serializers.ValidationError(["Jinja template is incorrect"])
         elif filtering_term_type == ChannelFilter.FILTERING_TERM_TYPE_REGEX or filtering_term_type is None:
             if filtering_term is not None:
                 if not is_regex_valid(filtering_term):
                     raise serializers.ValidationError(["Regular expression is incorrect"])
         else:
-            raise serializers.ValidationError([f"Expression type is incorrect"])
+            raise serializers.ValidationError(["Expression type is incorrect"])
         return data
 
 
