@@ -1,6 +1,6 @@
 import json
 
-from apps.alerts.paging import check_user_availability, direct_paging, unpage_user
+from apps.alerts.paging import DirectPagingAlertGroupResolvedError, check_user_availability, direct_paging, unpage_user
 from apps.slack.scenarios import scenario_step
 from apps.slack.scenarios.paging import (
     DIRECT_PAGING_SCHEDULE_SELECT_ID,
@@ -65,15 +65,19 @@ class ManageRespondersUserChange(scenario_step.ScenarioStep):
                 view=view,
             )
         else:
-            # no warnings, proceed with paging
-            direct_paging(
-                organization=organization,
-                team=alert_group.channel.team,
-                from_user=slack_user_identity.get_user(organization),
-                users=[(selected_user, False)],
-                alert_group=alert_group,
-            )
-            view = render_dialog(alert_group)
+            try:
+                # no warnings, proceed with paging
+                direct_paging(
+                    organization=organization,
+                    team=alert_group.channel.team,
+                    from_user=slack_user_identity.get_user(organization),
+                    users=[(selected_user, False)],
+                    alert_group=alert_group,
+                )
+                view = render_dialog(alert_group)
+            except DirectPagingAlertGroupResolvedError:
+                view = render_dialog(alert_group, alert_group_resolved_warning=True)
+
             self._slack_client.api_call(
                 "views.update",
                 trigger_id=payload["trigger_id"],
@@ -90,14 +94,18 @@ class ManageRespondersConfirmUserChange(scenario_step.ScenarioStep):
         selected_user = _get_selected_user_from_payload(payload)
         organization = alert_group.channel.organization
 
-        direct_paging(
-            organization=organization,
-            team=alert_group.channel.team,
-            from_user=slack_user_identity.get_user(organization),
-            users=[(selected_user, False)],
-            alert_group=alert_group,
-        )
-        view = render_dialog(alert_group)
+        try:
+            direct_paging(
+                organization=organization,
+                team=alert_group.channel.team,
+                from_user=slack_user_identity.get_user(organization),
+                users=[(selected_user, False)],
+                alert_group=alert_group,
+            )
+            view = render_dialog(alert_group)
+        except DirectPagingAlertGroupResolvedError:
+            view = render_dialog(alert_group, alert_group_resolved_warning=True)
+
         self._slack_client.api_call(
             "views.update",
             trigger_id=payload["trigger_id"],
@@ -114,17 +122,22 @@ class ManageRespondersScheduleChange(scenario_step.ScenarioStep):
         selected_schedule = _get_selected_schedule_from_payload(payload)
         organization = alert_group.channel.organization
 
-        direct_paging(
-            organization=organization,
-            team=alert_group.channel.team,
-            from_user=slack_user_identity.get_user(organization),
-            schedules=[(selected_schedule, False)],
-            alert_group=alert_group,
-        )
+        try:
+            direct_paging(
+                organization=organization,
+                team=alert_group.channel.team,
+                from_user=slack_user_identity.get_user(organization),
+                schedules=[(selected_schedule, False)],
+                alert_group=alert_group,
+            )
+            view = render_dialog(alert_group)
+        except DirectPagingAlertGroupResolvedError:
+            view = render_dialog(alert_group, alert_group_resolved_warning=True)
+
         self._slack_client.api_call(
             "views.update",
             trigger_id=payload["trigger_id"],
-            view=render_dialog(alert_group),
+            view=view,
             view_id=payload["view"]["id"],
         )
 
@@ -150,7 +163,7 @@ class ManageRespondersRemoveUser(scenario_step.ScenarioStep):
 # slack view/blocks rendering helpers
 
 
-def render_dialog(alert_group):
+def render_dialog(alert_group, alert_group_resolved_warning=False):
     blocks = []
 
     # Show list of users that are currently paged
@@ -170,6 +183,15 @@ def render_dialog(alert_group):
         ]
     if paged_users:
         blocks += [DIVIDER_BLOCK]
+
+    # Show a warning when trying to add responders for a resolved alert group
+    if alert_group_resolved_warning:
+        blocks += [
+            {
+                "type": "section",
+                "text": {"type": "mrkdwn", "text": f":no_entry: {DirectPagingAlertGroupResolvedError.DETAIL}"},
+            }
+        ]
 
     # Show user and schedule dropdowns
     input_id_prefix = _generate_input_id_prefix()
